refactor: replace preview schema with viewer contract
