+++ packages/db/prisma/schema.prisma
@@ -20,12 +20,6 @@ enum JobStatus {
   dead
 }
 
-enum PreviewStatus {
-  pending
-  ready
-  failed
-}
-
 enum ShareTargetType {
   file
   folder
@@ -144,7 +138,6 @@ model File {
   mimeType         String
   sizeBytes        BigInt
   contentChecksum  String?
-  previewStatus    PreviewStatus @default(pending)
   deletedAt        DateTime?
   createdAt        DateTime      @default(now())
   updatedAt        DateTime      @updatedAt
